[Images] Improve type stability of Image constructor
Inference is run before the inlining pass, so when inference runs on constructors of the form `T{ndims(A)}(...)`, it does't see the inlined constant; it just sees an Int.  The fix is to just use the type parameters directly.  I've added this as an additional constructor to make sure it doesn't break any existing code with very poorly typed arrays.