[ImageCore] Support colorview(ARGB32, ::Array{C<:Colorant}). Fixes #38.
This takes another step in the direction of discouraging explicit use of reinterpret. This is considerably more careful to check whether the eltype of input arrays matches the eltype of the Colorant before using reinterpret, and consequently is more reliable about falling back to ColorView. It resolves the alignment issue by avoiding calling reinterpret.

To simplify the code, this uses triangular dispatch, so this also bumps the minimum required version of Julia.