[ImageMorphology] Modernize `label_components` (#46)
`label_components` is one of the more important algorithms here,
but it was written to a very old version of Julia and not substantially
updated since. This reworks it to a more modern approach.
From the user perspective, the biggest change is the switch to a kwarg
interface for `dims`.

Because `imfill` calls `label_components`, it too was reworked similarly.