[ImageMorphology] `feature_transform`: add multithreading
If Julia is being run with multiple threads, this now will default
to using all threads in `feature_transform`. The multithreaded
implementation is relatively straightforward, although the recursive algorithm
makes it a little hard to wrap your head around:

- `computeft!` for dimension `d` is dependent only on slices of dimensions `1:d`,
  and indeed for all except the final call to `voronoift!` just on dimensions `1:d-1`.
  Hence you can divide the last dimension `N` into chunks, skip the final
  `voronoift!` on dimension `N`, and give each to a thread.

- For the final `voronoift!` call along dimension `N`, it's a
  one-dimensional operation along this dimension. Hence you can just
  split the next-to-last dimension into chunks instead.

This change was responsible for performance improvements in `distance_transform`
noted in https://github.com/JuliaImages/image_benchmarks/issues/1.

Julia 1.3 is required for `@spawn`

Co-authored-by: Johnny Chen <johnnychen94@hotmail.com>